Changed styling of pages
