+++ pages/impact-areas.js
@@ -1,18 +1,11 @@
-import Head from 'next/head';
 import Link from 'next/link';
 import styles from '../styles/ImpactAreas.module.css';
 import Header from '../components/Header';
 
 export default function ImpactAreas() {
 	return (
-		<div>
-			<Head>
-				<title>Inclusion Labs</title>
-				<meta name='description' content='Generated by create next app' />
-			</Head>
-
+		<>
 			<Header />
-
 			<main className={styles.main}>
 				<h1 className={styles.title}>Choose an impact area</h1>
 
@@ -72,6 +65,6 @@ export default function ImpactAreas() {
 					</Link>
 				</div>
 			</main>
-		</div>
+		</>
 	);
 }
